move bad versions to the repo
