+++ src/main/java/at/hannibal2/skyhanni/data/jsonobjects/repo/DiscontinuedMinecraftVersionsJson.kt
@@ -10,4 +10,5 @@ data class DiscontinuedMinecraftVersionsJson(
 data class DiscontinuedMinecraftVersion(
     @Expose @SerializedName("extra_info") val extraInfo: List<String>? = null,
     @Expose @SerializedName("config_info") val configInfo: List<String>? = null,
+    @Expose @SerializedName("luck_amount") val luckAmount: Float? = null,
 )
